constroller, api refactor

+++ app/Http/Controllers/Api/UsersController.php
@@ -22,10 +22,11 @@ class UsersController extends Controller
         return response()->json($result, 200);
     }
 
+    //create the user
     public function addUsers(Request $request) {
         {   
             $rules = array(
-                "id" => "string|integer",
+                "id" => "integer",
                 "name" => "string|required|max:50",
                 "email" => "string",
                 "password" => "required",
@@ -36,37 +37,27 @@ class UsersController extends Controller
                 "integer" => "that is not a integer",
                 "required" => "value is missing",
                 "max:50" => "number of characters exceeded"
-            ];
-
-
-            $isValid = Validator::make($request->all(), $rules, $messages);
-           
+            ];                    
 
             $user = new User;
             $user->id = $request->id;
             $user->name = $request->name;
             $user->email = $request->email;
             $user->password = $request->password;
-
             $result = $user->save();
 
             $isValid = Validator::make($request->all(), $rules, $messages);
-            if ($isValid->fails())
-            {
-                return response()->json($isValid->errors(), 400);
+
+            if ($isValid->fails())            {
+                return response()->json($isValid->errors(), 400);                
             } else {
-                return response()-json(["result"=>"user added successfully", "user" => $result], 201);
+                return response()->json(["result"=>"user added successfully", "user" => $user], 201);
             };
         }
 
     }
 
-    public function getSortedUsersByName()
-    {
-        return User::orderBy('name')
-            ->get();
-    }
-
+    //get the users by his ID
     public function getUsersById($id)
     {
         $result = User::select('name', 'email')
@@ -77,16 +68,18 @@ class UsersController extends Controller
             $result = "users doesnt exist";
             return response()->json($result,404);
         } else {        
-        return response()->json($result,200);
+            return response()->json($result,200)->header();
         };
         
         
     }
 
+    //update the user by his ID
     public function updateUsersById(Request $request, $id)
     {
        
-        $user = User::findOrFail($id);
+        $user = User::findOrFail($id); //findOrFail() automatically sends status code 404 if resource is not found
+        
 
         $user->name = $request->name;
         $user->email = $request->email;    
@@ -102,25 +95,32 @@ class UsersController extends Controller
             
     }
 
-
+    
+    //delete the user by his ID
     public function deleteUsersById($id)
     {
-        $user = User::findOrFail($id);  //automatically sends status code 404 if resource is not found
-        $user->delete();
-
-        return [ "result" => "user was deleted"];
+        $user = User::find($id);  
+        if(!$user){
+            return response()->json(["result" => "no such user"], 404);
+        } else {
+            $user->delete();
+            return response()->json([ "result" => "user was deleted"], 204);
+        };
     }
 
+    
+    //search the users by their name
     public function searchUserByName($name)
     {
-        $users = User::where('name', 'like' , '%'.$name.'%')->get('name');
-        $count = count($users);
+        $users = User::where('name', 'like' , '%'.$name.'%')->get('name');       
 
         if (count($users) !== 0){
 
             return response()->json( ["count" => $count, "users" => $users], 200);
+
         } else {
-            return response()->json(["result" => "no users found"], 404);
+
+            return response()->json(["result" => "no user has benn found"], 404);
         }
 
     }
